# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Node.js Express REST API",
      "version": "1.0.0",
      "description": "API para gestionar usuarios con autenticación JWT y encriptación de contraseñas"
    },
    "servers": [
      {
        "url": "http://localhost:3003",
        "description": "Servidor local"
      }
    ],
    "paths": {
      "/": {
        "get": {
          "summary": "Bienvenida",
          "description": "Mensaje de bienvenida",
          "responses": {
            "200": {
              "description": "Mensaje de bienvenida",
              "content": {
                "application/json": {
                  "example": {
                    "message": "¡Bienvenido a Node.js Express REST API!"
                  }
                }
              }
            }
          }
        }
      },
      "/token": {
        "post": {
          "summary": "Generar Token",
          "description": "Genera un token JWT",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "user123"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Token generado",
              "content": {
                "text/plain": {
                  "example": "Token generado: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "summary": "Obtener Usuario por ID",
          "description": "Obtiene un usuario por su ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Usuario encontrado",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "nombre": {
                        "type": "string",
                        "example": "John"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Token inválido"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        },
        "put": {
          "summary": "Actualizar Usuario",
          "description": "Actualiza un usuario por su ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "example": "John"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Usuario actualizado correctamente"
            },
            "401": {
              "description": "Token inválido"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        },
        "delete": {
          "summary": "Eliminar Usuario",
          "description": "Elimina un usuario por su ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Usuario eliminado correctamente"
            },
            "401": {
              "description": "Token inválido"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        }
      },
      "/users": {
        "get": {
          "summary": "Obtener Todos los Usuarios",
          "description": "Obtiene todos los usuarios",
          "responses": {
            "200": {
              "description": "Lista de usuarios",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "nombre": {
                          "type": "string",
                          "example": "John"
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Token no proporcionado"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        },
        "post": {
          "summary": "Agregar Usuario",
          "description": "Agrega un nuevo usuario",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "example": "John"
                    },
                    "contrasea": {
                      "type": "string",
                      "example": "password123"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Usuario agregado correctamente"
            },
            "401": {
              "description": "Token no proporcionado"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "Iniciar Sesión",
          "description": "Inicia sesión y genera un token JWT",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "username": {
                      "type": "string",
                      "example": "john_doe"
                    },
                    "password": {
                      "type": "string",
                      "example": "password123"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Inicio de sesión exitoso",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Credenciales inválidas"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        }
      },
      "/login/reset-password/{id}": {
        "post": {
          "summary": "Resetear Contraseña",
          "description": "Resetear la contraseña del usuario",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "newPassword": {
                      "type": "string",
                      "example": "newPassword123"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Contraseña del usuario actualizada correctamente"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        }
      },
      "/users/confirm-password/{id}": {
        "post": {
          "summary": "Confirmar Contraseña",
          "description": "Confirmar la contraseña antes de editar un usuario",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID del usuario",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "password": {
                      "type": "string",
                      "example": "password123"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Contraseña correcta"
            },
            "401": {
              "description": "Contraseña incorrecta o token inválido"
            },
            "404": {
              "description": "Usuario no encontrado"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        }
      },
      "/users/hashs": {
        "post": {
          "summary": "Agregar Usuario con Contraseñas Hasheadas",
          "description": "Agrega un nuevo usuario con contraseñas hasheadas en diferentes formatos",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "nombre": {
                      "type": "string",
                      "example": "John"
                    },
                    "contrasea": {
                      "type": "string",
                      "example": "password123"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Usuario agregado correctamente"
            },
            "401": {
              "description": "Token no proporcionado o inválido"
            },
            "500": {
              "description": "Error interno del servidor"
            }
          }
        }
      }
    }
  }